backend/requirements.txt
Flask==2.3.2
flask-mysqldb==1.0.1
mysql-connector-python==8.0.33
